modify uriCheck,modify gas

+++ config/config.go
@@ -45,11 +45,11 @@ type (
 		ProjectKey       string `mapstructure:"project_key"`
 		ChainAccountAddr string `mapstructure:"chain_account_addr"`
 
-		GasCoefficient uint64 `mapstructure:"gas_coefficient"`
-		Gas            uint64 `mapstructure:"gas"`
-		Denom          string `mapstructure:"denom"`
-		Amount         int64  `mapstructure:"amount"`
-		AccoutGas      int64  `mapstructure:"account_gas"`
+		GasCoefficient float64 `mapstructure:"gas_coefficient"`
+		Gas            uint64  `mapstructure:"gas"`
+		Denom          string  `mapstructure:"denom"`
+		Amount         int64   `mapstructure:"amount"`
+		AccoutGas      int64   `mapstructure:"account_gas"`
 	}
 
 	// Server define a struct for http server
